Extract logged in/out layouts to own components
